cleared warnings.

+++ IGA.cpp
@@ -11,34 +11,34 @@
 
 
 IGA::IGA(float step_size_) {
-    step_size = step_size_;
+    step_size = static_cast<double>(step_size_);
 
     srand(static_cast<unsigned int>(time(nullptr))); 
-    policy[0] = static_cast<float>(rand()) / RAND_MAX;
+    policy[0] = rand() / RAND_MAX;
 
     srand(static_cast<unsigned int>(time(nullptr))); 
-    policy[1] = static_cast<float>(rand()) / RAND_MAX;
+    policy[1] = rand() / RAND_MAX;
 
     srand(static_cast<unsigned int>(time(nullptr))); 
-    policy[2] = static_cast<float>(rand()) / RAND_MAX;
+    policy[2] = rand() / RAND_MAX;
 
     // normalize
-    float sum = 0;
+    double sum = 0;
     sum = policy[0] + policy[1] + policy[2];
 
     for (unsigned long i=0; i<3; i++) {
-        policy[i] = static_cast<float>(policy[i])/sum;
+        policy[i] = policy[i]/sum;
     }
 }
 
 // returns greedy action according to policy.
 int IGA::greedy(Strategy strat) {
     srand(static_cast<unsigned int>(time(nullptr))); 
-    float random = static_cast<float>(rand()) / RAND_MAX;
+    double random = rand() / RAND_MAX;
 
     // choose action based on policy probabilities.
-    float p_rock = strat[0];
-    float p_paper = strat[1];
+    double p_rock = strat[0];
+    double p_paper = strat[1];
 
     if (random < p_rock) {
         int action = 0; // rock.
@@ -69,20 +69,20 @@ void IGA::observe(Reward r, Strategy x, Action action_y, Strategy true_y) {
     RockPaperScissors RPS;
 
     // derivative for probability of action 0:
-    float r_00 = RPS.step(0, 0)[0];
-    float r_01 = RPS.step(0, 1)[0];
-    float r_02 = RPS.step(0, 2)[0];
-    float d_p0 = r_00*true_y[0] + r_01*true_y[1] + r_02*true_y[2];
+    double r_00 = RPS.step(0, 0)[0];
+    double r_01 = RPS.step(0, 1)[0];
+    double r_02 = RPS.step(0, 2)[0];
+    double d_p0 = r_00*true_y[0] + r_01*true_y[1] + r_02*true_y[2];
     // derivative for probability of action 1:
-    float r_10 = RPS.step(1, 0)[0];
-    float r_11 = RPS.step(1, 1)[0];
-    float r_12 = RPS.step(1, 2)[0];
-    float d_p1 = r_10*true_y[0] + r_11*true_y[1] + r_12*true_y[2];
+    double r_10 = RPS.step(1, 0)[0];
+    double r_11 = RPS.step(1, 1)[0];
+    double r_12 = RPS.step(1, 2)[0];
+    double d_p1 = r_10*true_y[0] + r_11*true_y[1] + r_12*true_y[2];
     // derivative for probability of action 2:
-    float r_20 = RPS.step(2, 0)[0];
-    float r_21 = RPS.step(2, 1)[0];
-    float r_22 = RPS.step(2, 2)[0];
-    float d_p2 = r_20*true_y[0] + r_21*true_y[1] + r_22*true_y[2];
+    double r_20 = RPS.step(2, 0)[0];
+    double r_21 = RPS.step(2, 1)[0];
+    double r_22 = RPS.step(2, 2)[0];
+    double d_p2 = r_20*true_y[0] + r_21*true_y[1] + r_22*true_y[2];
     
     // update policy
     policy[0] += step_size * d_p0;
@@ -96,10 +96,10 @@ void IGA::observe(Reward r, Strategy x, Action action_y, Strategy true_y) {
     }
 
     // normalize policy
-    float sum = 0;
+    double sum = 0;
     sum = policy[0] + policy[1] + policy[2];
 
     for (unsigned long i=0; i<3; i++) {
-        policy[i] = static_cast<float>(policy[i])/sum;
+        policy[i] = policy[i]/sum;
     }
 }
